update unit tests, tweak Pico sound

+++ src/pico_pwm_audio_sink.cpp
@@ -72,8 +72,8 @@ void PicoPwmAudioSink::core1Task() {
     while (sink->running) {
         queue_remove_blocking(&sink->sampleQueue, buffer.data());
         for (size_t i = 0; i < sink->bufferSize && sink->running; ++i) {
-            const float scaled = buffer[i] / 32768.0f;    // [-1, +1)
-            const float biased = (scaled + 1.0f) * 0.5f;  // [0, 1)
+            const float scaled = static_cast<float>(buffer[i]) / 32768.0f;
+            const float biased = (scaled + 1.0f) * 0.5f;
             const uint16_t pwm_val = std::clamp<int>(biased * pwm_max, 0, pwm_max);
             pwm_set_gpio_level(sink->gpioPin, pwm_val);
             sleep_us(static_cast<uint32_t>(us_per_sample));
